requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.3
weasyprint>=62.0
pydantic>=2.7.4
python-dotenv>=1.0.1
requests>=2.32.3

# LangChain & LangGraph stack
langchain>=0.2.5
langchain-core>=0.2.5
langchain-openai>=0.1.6
langchain-community>=0.2.5
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.2.1
langgraph>=0.2.13

# Model providers & search
openai>=1.35.7
tavily-python>=0.3.7
duckduckgo-search>=5.3.0
ddgs

# Vector store & embeddings
faiss-cpu>=1.7.4
huggingface-hub>=0.23.4
sentence-transformers>=3.0.1
